Fixed errors and added delete alert

+++ src/app/components/comment-list/comment-list.component.ts
@@ -3,6 +3,8 @@ import { Subscription } from 'rxjs';
 import { Comment } from '../../shared/models/comment.model';
 import { CommentService } from '../../shared/service/comment.service';
 import { UserService } from '../user/user.service';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
+import { MatDialog } from '@angular/material';
 
 /**
  * Comment-list component gets a list of comments from the database.
@@ -27,11 +29,16 @@ export class CommentListComponent implements OnInit, OnDestroy {
   userId: string;
   userIsAuthenticated = false;
   private authStatusSub: Subscription;
+  title = 'confirmation-dialog';
 
   /**
    * @ignore
    */
-  constructor(public commentService: CommentService, private userService: UserService) { }
+  constructor(
+    public commentService: CommentService,
+    private userService: UserService,
+    public dialog: MatDialog
+    ) { }
 
   /**
    * This function performs a GET request from the CommentService for a list of comments from the database.
@@ -52,11 +59,19 @@ export class CommentListComponent implements OnInit, OnDestroy {
       });
   }
   /**
-   * Performs a delete function from the CommentService on a button click.
-   * @param commentId of type string.
+   * Opens a dialog popup when the delete button is clicked
+   * @param commentId string
    */
-  onDelete(commentId: string) {
-    this.commentService.deleteComment(commentId);
+  openDialog(commentId: string) {
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '350px',
+      data: 'Are you sure you want this deleted?'
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.commentService.deleteComment(commentId);
+      }
+    });
   }
 
   /**
